Refactor contract CLI tool externally
